Refactor exception handling
